refactor: Move optional tests to their own file
